requirements.txt
appnope==0.1.0
beautifulsoup4==4.4.1
click==5.1
decorator==4.0.4
flake8==2.4.1
Flask==0.10.1
gnureadline==6.3.3
gunicorn==19.3.0
ipython==4.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mccabe==0.3.1
path.py==8.1.1
pep8==1.5.7
pep8-naming==0.3.3
pexpect==3.3
pickleshare==0.5
py==1.4.30
pyflakes==0.8.1
pytest==2.8.0
python-dotenv==0.1.3
PyYAML==3.11
requests==2.7.0
simplegeneric==0.8.1
traitlets==4.0.0
Werkzeug==0.10.4
wheel==0.24.0
